requirements.txt
azure-core==1.29.7
azure-storage-blob==12.19.0
bcrypt==4.1.2
behave==1.2.6
fastapi==0.109.2
jose==1.0.0
jwt==1.3.1
passlib==1.7.4
psycopg2==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
SQLAlchemy==2.0.25
urllib3==2.1.0
uvicorn==0.27.0

